# packages/shared/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/support/tickets:
    post:
      operationId: SupportController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketDto"
      responses:
        "201":
          description: ""
      security: &a25
        - bearer: []
      summary: Open a support ticket
      tags: &a26
        - support
    get:
      operationId: SupportController_list
      parameters: []
      responses:
        "200":
          description: ""
      security: *a25
      summary: List tickets (own for users; all for staff/admin)
      tags: *a26
  /api/v1/support/tickets/{id}:
    get:
      operationId: SupportController_getOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a25
      summary: Ticket detail with its conversation
      tags: *a26
    patch:
      operationId: SupportController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTicketDto"
      responses:
        "200":
          description: ""
      security: *a25
      summary: Update ticket status/priority/assignee (staff/admin)
      tags: *a26
  /api/v1/support/tickets/{id}/reply:
    post:
      operationId: SupportController_reply
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplyTicketDto"
      responses:
        "201":
          description: ""
      security: *a25
      summary: Reply on the ticket conversation
      tags: *a26
...
info:
  title: EGT B2B API
  description: Eagle Goods Trading Co. — sourcing & export backend API (v1)
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateTicketDto:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
        priority:
          type: string
          enum:
            - low
            - medium
            - high
            - urgent
          default: medium
      required:
        - subject
        - body
    ReplyTicketDto:
      type: object
      properties:
        body:
          type: string
      required:
        - body
    UpdateTicketDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - open
            - in_progress
            - waiting_on_customer
            - resolved
            - closed
        priority:
          type: string
          enum:
            - low
            - medium
            - high
            - urgent
        assigneeId:
          type: string
...